refactor: src/common/ui -> src/common/components
Components includes non-UI components, so I moved them
