+++ src/common/components/atoms/avatar.tsx
