+++ src/common/components/atoms/qr-code.tsx
