+++ src/common/lib/theme/ThemeEditorToolbar.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect } from "react";
-import { Card, CardFooter } from "@/common/ui/atoms/card";
+import { Card, CardFooter } from "@/common/components/atoms/card";
 import { FaFloppyDisk } from "react-icons/fa6";
 import { ThemeSettings } from "@/common/lib/theme";
 import { Color, FontFamily } from "@/common/lib/theme";
 import DEFAULT_THEME from "@/common/lib/theme/defaultTheme";
-import ColorSelector from "@/common/ui/molecules/ColorSelector";
-import FontSelector from "@/common/ui/molecules/FontSelector";
+import ColorSelector from "@/common/components/molecules/ColorSelector";
+import FontSelector from "@/common/components/molecules/FontSelector";
 
 export type ThemeEditorToolbarArgs = {
   theme: ThemeSettings;
